check if should reconcile before reconciling for all resources

+++ controllers/application/authorization_policy.go
@@ -22,6 +22,17 @@ func (r *ApplicationReconciler) reconcileAuthorizationPolicy(ctx context.Context
 	}
 	defaultDenyAuthPolicy := getDefaultDenyPolicy(application, defaultDenyPaths)
 
+	shouldReconcile, err := r.ShouldReconcile(ctx, &defaultDenyAuthPolicy)
+	if err != nil {
+		r.SetControllerFinishedOutcome(ctx, application, controllerName, err)
+		return reconcile.Result{}, err
+	}
+
+	if !shouldReconcile {
+		r.SetControllerFinishedOutcome(ctx, application, controllerName, err)
+		return reconcile.Result{}, nil
+	}
+
 	if application.Spec.AuthorizationSettings != nil {
 		if application.Spec.AuthorizationSettings.AllowAll == true {
 			err := r.GetClient().Delete(ctx, &defaultDenyAuthPolicy)
@@ -36,7 +47,7 @@ func (r *ApplicationReconciler) reconcileAuthorizationPolicy(ctx context.Context
 		}
 	}
 
-	_, err := ctrlutil.CreateOrPatch(ctx, r.GetClient(), &defaultDenyAuthPolicy, func() error {
+	_, err = ctrlutil.CreateOrPatch(ctx, r.GetClient(), &defaultDenyAuthPolicy, func() error {
 		err := ctrlutil.SetControllerReference(application, &defaultDenyAuthPolicy, r.GetScheme())
 		if err != nil {
 			r.SetControllerError(ctx, application, controllerName, err)
